Update crud.py
implemented oops concept

+++ backend/crud.py
@@ -5,10 +5,25 @@ from config import mydb
 from flask import jsonify
 from flask import request
 from app import app
-from db_services import execute,closeConnection, cursor
-#insert audio details into audio table
+#creating a class named Audio
+class Audio:
+    def __init__(self, trackid: str, title: str, category: str, author: str, image: str):
+        self.trackid = trackid
+        self.title = title
+        self.category = category
+        self.author = author
+        self.image = image
+# creating a class named User
+class User:
+    def __init__(self, userid: str, fullname: str, username: str, password: str, usertype: str):
+        self.userid = userid
+        self.fullname = fullname
+        self.username = username
+        self.password = password
+        self.usertype = usertype
+# insert audio details into table
 @app.route('/insert', methods=['POST'])
-def createAudio():
+def createAudio(trackid=None):
     try:
         json = request.json
         print(json)
@@ -16,10 +31,14 @@ def createAudio():
         category = json['category']
         author = json['author']
         image = json['image']
+        audio = Audio(trackid, title, category, author, image)
         if title and category and author and image and request.method == 'POST':
-            sqlQuery = "INSERT INTO audio( title, category, author, image) VALUES( %s, %s, %s,%s)"
-            bindData = ( title, category, author, image)
-            execute(sqlQuery,bindData)
+            conn = mydb.connect()
+            cursor = conn.cursor(pymysql.cursors.DictCursor)
+            sqlQuery = "INSERT INTO audio(title, category, author, image) VALUES( %s, %s, %s,%s)"
+            bindData = (audio.title, audio.category, audio.author, audio.image)
+            cursor.execute(sqlQuery, bindData)
+            conn.commit()
             respone = jsonify('Audio added successfully!')
             respone.status_code = 200
             return respone
@@ -28,28 +47,34 @@ def createAudio():
     except Exception as e:
         print(e)
         return 'Exception'
-
-#view all datas in audio table
-@app.route('/view_audios', methods =['GET'])
+    finally:
+        cursor.close()
+        conn.close()
+# view all datas from audio table
+@app.route('/view_audios', methods=['GET'])
 def viewAudios():
     try:
         conn = mydb.connect()
         cursor = conn.cursor(pymysql.cursors.DictCursor)
         cursor.execute("SELECT * FROM audio")
         empRows = cursor.fetchall()
+        conn.commit()
         respone = jsonify(empRows)
         respone.status_code = 200
         return respone
-    except Exception as e: 
+    except Exception as e:
         print(e)
-     
-#view particular data from audio table
+# view particular data from table
 @app.route('/view_audio/<trackid>', methods=['GET'])
-def audioAetails(trackid):
+def audioDetails(trackid, title=None, category=None, author=None, image=None):
     try:
-        sqlQuery="SELECT trackid , title, category, author, image FROM audio WHERE trackid =%s"
-        bindData =(trackid)
-        execute(sqlQuery,bindData)
+        audio = Audio(trackid, title, category, author, image)
+        print(audio.trackid)
+        conn = mydb.connect()
+        cursor = conn.cursor(pymysql.cursors.DictCursor)
+        sqlQuery = "SELECT trackid , title, category, author, image FROM audio WHERE trackid =%s"
+        bindData = audio.trackid
+        cursor.execute(sqlQuery, bindData)
         empRow = cursor.fetchone()
         respone = jsonify(empRow)
         respone.status_code = 200
@@ -57,25 +82,27 @@ def audioAetails(trackid):
     except Exception as e:
         print(e)
     finally:
-        closeConnection()
-#update audio from audio table
+        cursor.close()
+        conn.close()
+# update audio details of a particular track which is specified by track id
 @app.route('/update/<trackid>', methods=['PUT'])
 def updateAudio(trackid):
     try:
         _json = request.json
         print(_json)
-        _track_id = trackid
-        _title = _json['title']
-        _category= _json['category']
-        _author = _json['author']
-        _image = _json['image']
-       
-        if  _title and _category and _author and  _image and request.method  == 'PUT':   
-            sqlQuery = ("UPDATE audio SET title= %s, category= %s, author= %s, image=%s  WHERE trackid=%s")
-            bindData = ( _title, _category, _author,  _image,_track_id )
+        new_track_id = trackid
+        new_title = _json['title']
+        new_category = _json['category']
+        new_author = _json['author']
+        new_image = _json['image']
+        audio = Audio(new_track_id, new_title, new_category, new_author, new_image)
+        print(audio.trackid)
+        if new_title and new_category and new_author and new_image and request.method == 'PUT':
+            sqlQuery = " UPDATE audio SET title= %s, category= %s, author= %s, image=%s  WHERE trackid=%s "
+            bindData = (audio.title, audio.category, audio.author, audio.image, audio.trackid)
             conn = mydb.connect()
             cursor = conn.cursor()
-            cursor.execute(sqlQuery,bindData)
+            cursor.execute(sqlQuery, bindData)
             conn.commit()
             respone = jsonify('Audio updated successfully!')
             respone.status_code = 200
@@ -85,49 +112,60 @@ def updateAudio(trackid):
             return showMessage()
     except Exception as e:
         print(e)
-    
-#delete track from audio table
+# delete particular track
 @app.route('/delete/<trackid>', methods=['DELETE'])
-def deleteAudio(trackid):
+def deleteAudio(trackid, title=None, category=None, author=None, image=None):
     try:
-        sqlQuery="DELETE FROM audio WHERE trackid =%s"
-        bindData=(trackid)
-        execute(sqlQuery,bindData)
-        respone = jsonify('Audio deleted successfully!')
-        respone.status_code = 200
-        return respone
+        audio = Audio(trackid, title, category, author, image)
+        conn = mydb.connect()
+        cursor = conn.cursor(pymysql.cursors.DictCursor)
+        sqlQuery = "SELECT title FROM audio WHERE trackid =%s"
+        bindData = audio.trackid
+        data = cursor.execute(sqlQuery, bindData)
+        print(data)
+        if data == 0:
+            response = jsonify('Audio does not exist')
+            return response
+        elif data == 1:
+            sqlQuery = "DELETE FROM audio WHERE trackid =%s"
+            bindData = audio.trackid
+            cursor.execute(sqlQuery, bindData)
+            conn.commit()
+            respone = jsonify('Audio deleted successfully!')
+            respone.status_code = 200
+            return respone
     except Exception as e:
         print(e)
-
-# signup for user , with email, username and password and datas are enetered to user table
+# # signup for user , with email, username and password and datas are entered to user table
 @app.route('/register', methods=['POST'])
-def register():
+def register(userid=None):
     try:
         json = request.json
         print(json)
-        email= json['email']
+        fullname = json['fullname']
         username = json['username']
         password = json['password']
-        usertype = "user"
-        if email and username and password and usertype and request.method == 'POST':
+        usertype = 'user'
+        user = User(userid, fullname, username, password, usertype)
+        if fullname and username and password and usertype and request.method == 'POST':
             conn = mydb.connect()
             cursor = conn.cursor(pymysql.cursors.DictCursor)
-            query= "SELECT * FROM user WHERE username= '%s'" % (username)
-            data=cursor.execute(query)
+            query = "SELECT * FROM user WHERE username= %s"
+            bindData = user.username
+            data = cursor.execute(query, bindData)
             print(data)
-            if data>0:
-                conn.commit()
-                response = jsonify('User already Exsist!!')
+            if data > 0:
+                response = jsonify('User already Exist!!')
                 response.status_code = 200
                 return response
             else:
-                sqlQuery = "INSERT INTO user(email,username,password, usertype) VALUES(%s, %s, %s , %s)"
-                bindData = (email,username,password, usertype)
+                sqlQuery = "INSERT INTO user(fullname,username,password, usertype) VALUES(%s, %s, %s , %s)"
+                bindData = (user.fullname, user.username, user.password, user.usertype)
                 cursor.execute(sqlQuery, bindData)
                 conn.commit()
                 respone = jsonify('User added successfully!')
                 respone.status_code = 200
-                return respone      
+                return respone
         else:
             return showMessage()
     except Exception as e:
@@ -136,29 +174,31 @@ def register():
     finally:
         cursor.close()
         conn.close()
-#Login function , which enable the user to login to the application using username and password
+# Login function , which enable the user to login to the application using username and password
 @app.route('/login', methods=['POST'])
-def login():
+def login(userid=None, fullname=None, usertype=None):
     try:
         json = request.json
         print(json)
         username = json['username']
         password = json['password']
+        user = User(userid, fullname, username, password, usertype)
         print(username)
         if username and password and request.method == 'POST':
             conn = mydb.connect()
             cursor = conn.cursor(pymysql.cursors.DictCursor)
-            sqlQuery="SELECT * FROM user WHERE username= '%s'  and password='%s'" % (username, password)
-            data=cursor.execute(sqlQuery)
-            print(data)      
-            if data==1:
+            sqlQuery = "SELECT usertype FROM user WHERE username= '%s'  and password='%s'" % (user.username, user.password)
+            data = cursor.execute(sqlQuery)
+            print(data)
+            if data == 1:
                 row = cursor.fetchone()
-                usertype=row.get('usertype') 
+                print(row)
+                usertype = row.get('usertype')
                 access_token = jwt.encode({
                     'username': username
                 }, app.config['JWT_SECRET_KEY'])
                 conn.commit()
-                return jsonify(message='Login Successful', access_token=access_token ,usertype=usertype),200
+                return jsonify(message='Login Successful', access_token=access_token, usertype=usertype), 200
             else:
                 conn.commit()
                 return jsonify('Bad email or Password... Access Denied!'), 401
@@ -169,8 +209,7 @@ def login():
         return 'Exception'
     finally:
         cursor.close()
-        conn.close()  
-#shows the exeption       
+        conn.close()
 @app.errorhandler(404)
 def showMessage(error=None):
     message = {
@@ -180,9 +219,6 @@ def showMessage(error=None):
     respone = jsonify(message)
     respone.status_code = 404
     return respone
-
-
-closeConnection()
-
+#running the app
 if __name__ == "__main__":
     app.run()
